requirements.txt
matplotlib
pickle
antlr4-python3-runtime==4.7.2
networkx==2.4
python-telegram-bot==12.2.0
